refactor(GN-26): useReservationCalendar 추가
